+++ src/main/java/net/soulsweaponry/entity/mobs/ChaosMonarch.java
@@ -89,7 +89,7 @@ public class ChaosMonarch extends BossEntity implements GeoEntity {
         .add(EntityAttributes.GENERIC_MOVEMENT_SPEED, 0.15D)
         .add(EntityAttributes.GENERIC_ATTACK_DAMAGE, 20.0D)
         .add(EntityAttributes.GENERIC_KNOCKBACK_RESISTANCE, 1.0D)
-        .add(EntityAttributes.GENERIC_ARMOR, 4.0D)
+        .add(EntityAttributes.GENERIC_ARMOR, ConfigConstructor.chaos_monarch_armor)
         .add(EntityAttributes.GENERIC_ATTACK_KNOCKBACK, 2.0D);
     }