+++ src/main/java/net/soulsweaponry/entity/mobs/EvilForlorn.java
@@ -39,7 +39,8 @@ public class EvilForlorn extends Forlorn {
     public static DefaultAttributeContainer.Builder createForlornAttributes() {
         return MobEntity.createMobAttributes()
         .add(EntityAttributes.GENERIC_FOLLOW_RANGE, 250D)
-        .add(EntityAttributes.GENERIC_MAX_HEALTH, 15D)
+        .add(EntityAttributes.GENERIC_MAX_HEALTH, ConfigConstructor.evil_forlorn_health)
+        .add(EntityAttributes.GENERIC_ARMOR, ConfigConstructor.evil_forlorn_bonus_armor)
         .add(EntityAttributes.GENERIC_MOVEMENT_SPEED, 0.3000000003D)
         .add(EntityAttributes.GENERIC_ATTACK_DAMAGE, 2.0D);
     }